server/pyproject.toml
[project]
name = "air-dict-server"
version = "1.0.0"
description = "轻量级英文词典 API - 双向翻译 + 收藏功能"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "deep-translator>=1.11.4",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.1",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []
